guild.py update and some (?) other

+++ belphegor/remind.py
@@ -126,7 +126,8 @@ class Remind:
                 "timeout":  "Timeout, cancelled deleting."
             }
             check = await ctx.yes_no_prompt(sentences)
-            await self.event_list.delete_one(self_events[position-1])
+            if check:
+                await self.event_list.delete_one(self_events[position-1])
         else:
             await ctx.send("Position out of range.")
